Reworking architectrure
The database design was fundamentally flawed.
It was designed to allow compeltely synchronous access,
but this proved difficult in maintaining multiple open pages,
without the necessary systems in place for handling mutliuser scenarios,
a core feature from the beginning.

The database is now being rewritten in a thread-safe manner, and
 synchronous usage will be implemented on top of this via blocking calls
and optionally asynchronous structures.